Rewrote interface between Problem and optimization algorithms
Now all implemented optimization algorithms utilize the problem traits
in order to interface with arbitrary objective functions. This gives the
freedom to use any number system for the objective function without
interfering on the implementation of the optimization algorithms. Also,
rewrote the objective functions in examples and tests to utilize these
traits. Now the objective functions written so far utilize DualScalars
which improved performance dramatically (the benches show that the
solving time is about one half of what it used to be).